feat: Change IsNotEmptyString to IsStringWithValue
IsStringWithValue returns true if given value is a string and it is not empty

Closes #5

IsNotEmptyString was a bit weird in that it returned true for null and undefined.
While logically true it is not the commonn use case.  More often the user of the library would like to know if a string has value.